Created a new descartes_tests package and moved everything into that directory.
