+++ descartes_tests/test/moveit/launch/moveit_state_adapter.launch
@@ -6,7 +6,7 @@
   <arg name="robot_description" default="robot_description"/>
 
   <!-- URDF file (including path) -->
-  <arg name="urdf_file" default="$(find descartes_moveit)/test/resources/kuka_kr210/kr210l150.urdf"/>
+  <arg name="urdf_file" default="$(find descartes_tests)/test/moveit/resources/kuka_kr210/kr210l150.urdf"/>
 
   <!-- Launch test node -->
   <arg name="run_test_node" default="true"/>
@@ -15,16 +15,16 @@
   <param name="$(arg robot_description)" textfile="$(arg urdf_file)"/>
 
   <!-- The semantic description that corresponds to the URDF -->
-  <param name="$(arg robot_description)_semantic" textfile="$(find descartes_moveit)/test/resources/kuka_kr210/kuka_kr210.srdf" />
+  <param name="$(arg robot_description)_semantic" textfile="$(find descartes_tests)/test/moveit/resources/kuka_kr210/kuka_kr210.srdf" />
   
   <!-- Load updated joint limits (override information from URDF) -->
   <group ns="$(arg robot_description)_planning">
-    <rosparam command="load" file="$(find descartes_moveit)/test/resources/kuka_kr210/joint_limits.yaml"/>
+    <rosparam command="load" file="$(find descartes_tests)/test/moveit/resources/kuka_kr210/joint_limits.yaml"/>
   </group>
 
   <!-- Load default settings for kinematics; these settings are overridden by settings in a node's namespace -->
   <group ns="$(arg robot_description)_kinematics">
-    <rosparam command="load" file="$(find descartes_moveit)/test/resources/kuka_kr210/kinematics.yaml"/>
+    <rosparam command="load" file="$(find descartes_tests)/test/moveit/resources/kuka_kr210/kinematics.yaml"/>
   </group>
   
 </launch>
