+++ descartes_tests/test/moveit/launch/utest.launch
@@ -9,13 +9,13 @@
 
 
 
-  <include file="$(find descartes_moveit)/test/launch/moveit_state_adapter.launch">
+  <include file="$(find descartes_tests)/test/moveit/launch/moveit_state_adapter.launch">
     <arg name="robot_description" value="$(arg robot_description)"/>
-    <arg name="urdf_file" value="$(find descartes_moveit)/test/resources/kuka_kr210/kr210l150.urdf"/>
+    <arg name="urdf_file" value="$(find descartes_tests)/test/moveit/resources/kuka_kr210/kr210l150.urdf"/>
     <arg name="run_test_node" value="$(arg run_test_node)"/>
   </include>
 
   <!-- Test Node -->
-  <test test-name="utest" pkg="descartes_moveit" type="descartes_moveit_utest" if="$(arg run_test_node)"/>
+  <test test-name="utest" pkg="descartes_tests" type="descartes_tests_moveit_utest" if="$(arg run_test_node)"/>
   
 </launch>
