+++ descartes_tests/test/moveit/resources/kuka_kr210/kinematics.yaml
