+++ descartes_tests/test/moveit/resources/kuka_kr210/kuka_kr210.srdf
