+++ descartes_trajectory/CMakeLists.txt
@@ -24,7 +24,6 @@ endif()
 catkin_package(
   INCLUDE_DIRS
     include
-    test
   LIBRARIES
     descartes_trajectory
   CATKIN_DEPENDS
@@ -51,7 +50,6 @@ add_library(descartes_trajectory
             src/joint_trajectory_pt.cpp
             src/trajectory.cpp
             src/axial_symmetric_pt.cpp
-            test/cartesian_robot.cpp
 )
 
 target_link_libraries(descartes_trajectory
@@ -77,16 +75,11 @@ install(
 #############
 
 if(CATKIN_ENABLE_TESTING)
-  set(UTEST_SRC_FILES test/utest.cpp
-      test/trajectory_pt.cpp
-      test/cart_trajectory_pt.cpp
-      test/joint_trajectory_pt.cpp
-      test/cartesian_robot.cpp
-      test/cartesian_robot_test.cpp
-      test/axial_symmetric_pt.cpp)
-
+  set(UTEST_SRC_FILES
+    test/utest.cpp
+    test/trajectory_pt.cpp
+  )
   catkin_add_gtest(${PROJECT_NAME}_utest ${UTEST_SRC_FILES})
   target_compile_definitions(${PROJECT_NAME}_utest PUBLIC GTEST_USE_OWN_TR1_TUPLE=0)
   target_link_libraries(${PROJECT_NAME}_utest descartes_trajectory)
-
 endif()
